Data.IP

Description

Data structures to express IPv4, IPv6 and IP range.

IP data

data IP

A unified IP data for IP and IP. To create this, use the data constructors. Or use read "192.0.2.1" :: IP, for example. Also, "192.0.2.1" can be used as literal with OverloadedStrings.

>>> (read "192.0.2.1" :: IP) == IPv4 (read "192.0.2.1" :: IPv4)
True
>>> (read "2001:db8:00:00:00:00:00:01" :: IP) == IPv6 (read "2001:db8:00:00:00:00:00:01" :: IPv6)
True

Constructors

IPv4

Fields

ipv4 :: IPv4

IPv6

Fields

ipv6 :: IPv6

IPv4

data IPv4

The abstract data type to express an IPv4 address. To create this, use toIPv4. Or use read "192.0.2.1" :: IP, for example. Also, "192.0.2.1" can be used as literal with OverloadedStrings.

>>> read "192.0.2.1" :: IPv4
192.0.2.1

toIPv4 :: [Int] -> IPv4

The toIPv4 function returns the IP address corresponding to the given list of Int octets. The function is strict in the four elements of the list. An error is returned if the list has a differnet length. The input elements are silently truncated to their 8 least-significant bits before they are combined to form the IPv4 address.

>>> toIPv4 [192,0,2,1]
192.0.2.1

toIPv4w :: Word32 -> IPv4

The toIPv4w function constructs the IP address corresponding to the given Word32 value. Unlike the fromHostAddress function, it is strict in the input value, which here is in host byte order.

>>> toIPv4w 0xc0000201
192.0.2.1

Since: 1.7.9

fromIPv4 :: IPv4 -> [Int]

The fromIPv4 function returns the list of four Int octets corresponding to the given IP address.

>>> fromIPv4 (toIPv4 [192,0,2,1])
[192,0,2,1]

fromIPv4w :: IPv4 -> Word32

The fromIPv4w function returns a single Word32 value corresponding to the given the IP address. Unlike the toHostAddress function, the returned value is strictly evaluated, and is not converted to network byte order.

>>> fromIPv4w (toIPv4 [0xc0,0,2,1]) == 0xc0000201
True

Since: 1.7.9

fromHostAddress :: HostAddress -> IPv4

The fromHostAddress function converts HostAddress to IP.

toHostAddress :: IPv4 -> HostAddress

The toHostAddress function converts IP to HostAddress.

IPv6

data IPv6

The abstract data type to express an IPv6 address. To create this, use toIPv6. Or use read "2001:DB8::1" :: IP, for example. Also, "2001:DB8::1" can be used as literal with OverloadedStrings.

>>> read "2001:db8:00:00:00:00:00:01" :: IPv6
2001:db8::1
>>> read "2001:db8:11e:c00::101" :: IPv6
2001:db8:11e:c00::101
>>> read "2001:db8:11e:c00:aa:bb:192.0.2.1" :: IPv6
2001:db8:11e:c00:aa:bb:c000:201
>>> read "2001:db8::192.0.2.1" :: IPv6
2001:db8::c000:201
>>> read "0::ffff:192.0.2.1" :: IPv6
::ffff:192.0.2.1
>>> read "0::0:c000:201" :: IPv6
::192.0.2.1
>>> read "::0.0.0.1" :: IPv6
::1

toIPv6 :: [Int] -> IPv6

The toIPv6 function returns the IP address corresponding to the given list of eight 16-bit Ints. The function is strict in the eight elements of the list. An error is returned if the list has a differnet length. The input elements are in host byte order and are silently truncated to their 16 least-signicant bits before they are combined to form the IPv6 address.

>>> toIPv6 [0x2001,0xDB8,0,0,0,0,0,1]
2001:db8::1

toIPv6b :: [Int] -> IPv6

The toIPv6b function returns the IPv6 address corresponding to the given list of sixteen Int octets. The function is strict in the sixteen elements of the list. An error is returned if the list has a differnet length. The input elements are silently truncated to their 8 least-signicant bits before they are combined to form the IPv6 address.

>>> toIPv6b [0x20,0x01,0xD,0xB8,0,0,0,0,0,0,0,0,0,0,0,1]
2001:db8::1

toIPv6w :: (Word32, Word32, Word32, Word32) -> IPv6

The toIPv6w function constructs the IP address corresponding to the given four-tuple of host byte order Word32 values. This function differs from the fromHostAddress6 function only in the fact that it is strict in the elements of the tuple.

>>> toIPv6w (0x20010DB8,0x0,0x0,0x1)
2001:db8::1

Since: 1.7.9

fromIPv6 :: IPv6 -> [Int]

The fromIPv6 function returns a list eight Ints in host byte order corresponding to the eight 16-bit fragments of the given IPv6 address.

>>> fromIPv6 (toIPv6 [0x2001,0xDB8,0,0,0,0,0,1])
[8193,3512,0,0,0,0,0,1]

fromIPv6b :: IPv6 -> [Int]

The fromIPv6b function returns the 16 Int octets corresponding to the 16 bytes of the given IPv6 address.

>>> fromIPv6b (toIPv6b [0x20,0x01,0xD,0xB8,0,0,0,0,0,0,0,0,0,0,0,1])
[32,1,13,184,0,0,0,0,0,0,0,0,0,0,0,1]

fromIPv6w :: IPv6 -> (Word32, Word32, Word32, Word32)

The fromIPv6w function returns a four-tuple of Word32 values in host byte order corresponding to the given IP address. This is identical to the toHostAddress6 function, except that the elements of four-tuple are first strictly evaluated.

>>> fromIPv6w (toIPv6 [0x2001,0xDB8,0,0,0,0,0,1]) == (0x20010DB8, 0, 0, 1)
True

Since: 1.7.9

fromHostAddress6 :: HostAddress6 -> IPv6

The fromHostAddress6 function converts HostAddress6 to IP.

toHostAddress6 :: IPv6 -> HostAddress6

The toHostAddress6 function converts IP to HostAddress6.

Converters

ipv4ToIPv6 :: IPv4 -> IPv6

Convert IPv4 address to IPv4-embedded-in-IPv6

fromSockAddr :: SockAddr -> Maybe (IP, PortNumber)

Convert SockAddr to IP.

Since: 1.7.4.

toSockAddr :: (IP, PortNumber) -> SockAddr

Convert IP to SockAddr.

Since: 1.7.8.

IP range data

data IPRange

A unified data for AddrRange IP and AddrRange IP. To create this, use read "192.0.2.0/24" :: IPRange. Also, "192.0.2.0/24" can be used as literal with OverloadedStrings.

>>> (read "192.0.2.1/24" :: IPRange) == IPv4Range (read "192.0.2.0/24" :: AddrRange IPv4)
True
>>> (read "2001:db8:00:00:00:00:00:01/48" :: IPRange) == IPv6Range (read "2001:db8:00:00:00:00:00:01/48" :: AddrRange IPv6)
True

Constructors

IPv4Range

Fields

ipv4range :: AddrRange IPv4

IPv6Range

Fields

ipv6range :: AddrRange IPv6

data AddrRange a

The Addr range consists of an address, a contiguous mask, and mask length. The contiguous mask and the mask length are essentially same information but contained for pre calculation.

To create this, use makeAddrRange or read "192.0.2.0/24" :: AddrRange IP. Also, "192.0.2.0/24" can be used as literal with OverloadedStrings.

>>> read "192.0.2.1/24" :: AddrRange IPv4
192.0.2.0/24
>>> read "2001:db8:00:00:00:00:00:01/48" :: AddrRange IPv6
2001:db8::/48

Address class

class Eq a => Addr a where

>>> toIPv4 [127,0,2,1] `masked` intToMask 7
126.0.0.0

Methods

masked :: a -> a -> a

The masked function takes an Addr and a contiguous mask and returned a masked Addr.

intToMask :: Int -> a

The intToMask function takes an Int representing the number of bits to be set in the returned contiguous mask. When this integer is positive the bits will be starting from the MSB and from the LSB otherwise.

>>> intToMask 16 :: IPv4
255.255.0.0

>>> intToMask (-16) :: IPv4
0.0.255.255

>>> intToMask 16 :: IPv6
ffff::

>>> intToMask (-16) :: IPv6
::ffff

makeAddrRange :: Addr a => a -> Int -> AddrRange a

The makeAddrRange functions takes an Addr address and a mask length. It creates a bit mask from the mask length and masks the Addr address, then returns AddrRange made of them.

>>> makeAddrRange (toIPv4 [127,0,2,1]) 8
127.0.0.0/8
>>> makeAddrRange (toIPv6 [0x2001,0xDB8,0,0,0,0,0,1]) 8
2000::/8

(>:>) :: Addr a => AddrRange a -> AddrRange a -> Bool

The >:> operator takes two AddrRange. It returns True if the first AddrRange contains the second AddrRange. Otherwise, it returns False.

>>> makeAddrRange ("127.0.2.1" :: IPv4) 8 >:> makeAddrRange "127.0.2.1" 24
True
>>> makeAddrRange ("127.0.2.1" :: IPv4) 24 >:> makeAddrRange "127.0.2.1" 8
False
>>> makeAddrRange ("2001:DB8::1" :: IPv6) 16 >:> makeAddrRange "2001:DB8::1" 32
True
>>> makeAddrRange ("2001:DB8::1" :: IPv6) 32 >:> makeAddrRange "2001:DB8::1" 16
False

isMatchedTo :: Addr a => a -> AddrRange a -> Bool

The isMatchedTo function take an Addr address and an AddrRange, and returns True if the range contains the address.

>>> ("127.0.2.0" :: IPv4) `isMatchedTo` makeAddrRange "127.0.2.1" 24
True
>>> ("127.0.2.0" :: IPv4) `isMatchedTo` makeAddrRange "127.0.2.1" 32
False
>>> ("2001:DB8::1" :: IPv6) `isMatchedTo` makeAddrRange "2001:DB8::1" 32
True
>>> ("2001:DB8::" :: IPv6) `isMatchedTo` makeAddrRange "2001:DB8::1" 128
False

addrRangePair :: Addr a => AddrRange a -> (a, Int)

The unmakeAddrRange functions take a AddrRange and returns the network address and a mask length.

>>> addrRangePair ("127.0.0.0/8" :: AddrRange IPv4)
(127.0.0.0,8)
>>> addrRangePair ("2000::/8" :: AddrRange IPv6)
(2000::,8)

ipv4RangeToIPv6 :: AddrRange IPv4 -> AddrRange IPv6

Convert IPv4 range to IPV4-embedded-in-IPV6 range